app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-milvus
langchain-community==0.0.10
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.6
beautifulsoup4==4.12.2
